+++ src/main/java/br/com/sale/security/UserSS.java
@@ -13,12 +13,12 @@ import br.com.sale.domain.enums.Perfil;
 public class UserSS implements UserDetails {
 	private static final long serialVersionUID = 1L;
 
-	private Long id;
+	private Integer id;
 	private String email;
 	private String senha;
 	private Collection<? extends GrantedAuthority> authorities;
 
-	public UserSS(Long id, String email, String senha, Set<Perfil> perfis) {
+	public UserSS(Integer id, String email, String senha, Set<Perfil> perfis) {
 		this.id = id;
 		this.email = email;
 		this.senha = senha;
@@ -26,7 +26,7 @@ public class UserSS implements UserDetails {
 				.collect(Collectors.toList());
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
